Make workable in Rails 2.x
Rails 2 support isn't generally supported. But if you really want to make
it work you just need to mixin the module yourself with code that looks
like this in your environment.rb

    ActiveRecord::Base.extend Alphabar::AlphaScope
    ActionView::Base.send :include, Alphabar::Helper

There was one thing standing the way. The way the scope is generated is
slightly different between Rails 2 and 3. So for the sake letting someone
hack it in on a Rails 2 app I have added both methods of scope generation
depending on if "where" is available.